Refactored benchmarks
